Fix order of params on ProcessPool
On Códice Software we use more arguments that are not neccessary for this
use case. It is important for you to know that you can and MUST modify
the code of the agent to suit your needs! It is not written in stone,
and it is not as flexible as you might think.