WIP - Add netstandard versions of some assemblies
A Xamarin.Mac application can only depend on .NET Framework DLLs if they
are 4.0 or lower, which for some users would mean sacrifying the use of
new APIs. However, on Xamarin.Mac you can reference .NET Standard, so I
added a netstandard2.0 version of the following assemblies that the macOS
application relies on:

* lib
* guitestinterfaces

To prevent compiler from copying a lot of assemblies to the bin directory
when compiling these assemblies, we have to make sure that the application
that depends on them is .NET Framework +4.7.2, where the vast majority
of assemblies are the same across Framework and Standard.