Fix application startup code
Because we don't need any menu, I removed the MainMenu.xib file. As it turns out, that file was also in charge of pointing to the Application's delegate. You can manually set the delegate assigning an instance of it to the NSApplication.SharedApplication.Delegate property. You must do so only AFTER you have called NSApplication.Init(), otherwise the SharedApplication will be null.